FEAT: Adição das curvas e ranking

+++ src/main/java/projeto/apiData/services/EvasaoAgregada.java
@@ -2,20 +2,19 @@ package projeto.apiData.services;
 
 public class EvasaoAgregada {
 
-    private final long totalDesistentes;
+    private final Double totalDesistentes;
     private final long totalIngressantes;
 
-    public EvasaoAgregada(Long totalDesistentes, Long totalIngressantes) {
-        this.totalDesistentes = totalDesistentes != null ? totalDesistentes : 0;
+    public EvasaoAgregada(Double totalDesistentes, Long totalIngressantes) {
+        this.totalDesistentes = totalDesistentes != null ? totalDesistentes : 0.0;
         this.totalIngressantes = totalIngressantes != null ? totalIngressantes : 0;
     }
-
 
     public double getTaxaEvasao() {
         if (totalIngressantes == 0) return 0.0;
         return (double) totalDesistentes / totalIngressantes * 100;
     }
 
-    public long getTotalDesistentes() { return totalDesistentes; }
+    public Double getTotalDesistentes() { return totalDesistentes; }
     public long getTotalIngressantes() { return totalIngressantes; }
 }
